+++ CHANGES.md
@@ -8,6 +8,7 @@ Bug Fixes
 ---------
 * [#71](https://github.com/twall/jna/issues/71), [#73](https://github.com/twall/jna/issues/73): Fix OSGI entries in manifest - [twall](https://github.com/twall).
 * [#78](https://github.com/twall/jna/issues/78): Fix NPE in `platform.win32.Netapi32Util.getDomainTrusts` - [dblock](https://github.com/dblock).
+* Auto-sync memory for struct** arguments (array of struct pointers).
 
 Release 3.4.1
 =============